requirements.txt
Django==4.2.4
dj-database-url==2.1.0
django-ckeditor==6.7.0
django-crispy-forms==1.14.0
django-mathfilters==1.0.0
django-storages==1.14
gunicorn==23.0.0
Pillow==10.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
whitenoise==6.5.0
